MARS-519: code cleanup

+++ src/fdb5/fdb5_purge.cc
@@ -83,7 +83,7 @@ void FDBPurge::run()
 {
     eckit::option::CmdArgs args(&FDBPurge::usage, -1, options_);
 
-    for (int i = 0; i < args.count(); ++i) {
+    for (size_t i = 0; i < args.count(); ++i) {
 
         eckit::PathName path(args.args(i));
